Initial codes
make file not implemented
terminal input not implemented
(now need to hard code the file input path and output path)
in the program input width then height as instructed by the user prompts. 
select uniform or popularity partition for the image. 
there is a upper limit. image that are too big may not work.